bl2: Fix *_SZ define names
No functional changes. The *_SZ define names are confusing and sometimes
misleading. Fix that.

s/BL2IMG_TOTSZ/BL2IMG_SZ/: For the sake of consistency
s/BL2RAND_SZ/BL2IV_SZ/: Specify that is IV not just random bytes
s/BL2HDR_SZ/BL2BLKHDR_SZ/: This is the size of the block header, not of
  the whole BL2 (that is 4096 bytes)
s/BL2HASH_SZ/BL2SIG_SZ/: That's not an hash but a signature

Signed-off-by: Carlo Caione <ccaione@baylibre.com>