bl2: Extend signature functionality to G12A
The process of signing BL2 for the G12A SoCs is basically the same with
two major differences:

1) The G12A BL2 binary is bigger
2) On GXL the final size is obtained by cutting off a 4K tail to account
for the BL2 header.

This patch addresses the point (1) simply avoiding to hardcode the size
of the binary that instead is calculated at signing time. This is
causing all the *_SZ defines to be reworked to be more generic. For the
sake of clarity a new BL2HDR_SZ define is introduced carrying the size
of the whole BL2 header.

Point (2) is addressed simply looking at the size of the padded BL2
binary and accordingly fixing the output file size.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>