blx: Differentiate between sign and encrypt
While for the GXL family the BL3x payloads need to be encrypted, for AXG
and G12a they only need to be signed (at least when secure boot is off).
The original aml_encrypt tool uses two different parameters to
differentiate between these two cases: '--bl3enc' and '--bl3sig'.

To be able to support the G12a family in gxlimg we need to differentiate
between signing and encrypting using two new different parameters: '-s'
(for signing) and '-c' (for encrypting).

While we are at it change the stage name from bl3 to bl3x.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>