Add necessary to extract v3 FIPs
The V3 FIPs are generated for AXG and G12A/G12B/SM1 BootROMs.

The main difference is that the TOC is in clear and located right
after the BL2.
For G12A/G12B/SM1 the BL2 is larger (64K instead of 48K).

The V3 TOC also contains other entries for 1128bytes data,
this also adds the UUIDs for these and fixups to extract them.