Fix non-zero return code of gi_fip_extract_bl3x after success extraction
The ret would record the result of the last snprintf call, whether
it was successfull or not. Thus it would return a positive number
marking failure even after a successful extraction. And as that
would be reported as the program's return code, users would be
confused on whether the extraction was really failed.

Setting ret to 0 before returning to correctly reflect the result.