amlsblk: Don't assume topad is aligned to file block
gi_amlsblk_init() does choose padding to align to the block size, but
the size to align includes BL3xHDR_SZ. Thus hashsz is guaranteed to
not be aligned to the block size.

gi_amlsblk_flush_data had a bug that would fail to write padding if the
payload was aligned to the block size (gi_amlsblk_hash_payload was
fine). Both functions assumed that the last block of a file + topad was
less than the block size.

For example, in the previous code if `payloadsz = 512`, then the padding
would not be written. And if `payloadsz = 1000`, then uninitialized data
would be hashed and written:

```
==446768== Invalid read of size 16
==446768==    at 0x4D55B48: sha256_block_data_order_avx2 (sha256-x86_64.s:4223)
==446768==    by 0x4B83B5F: UnknownInlinedFun (md32_common.h:198)
==446768==    by 0x4B83B5F: SHA256_Update (md32_common.h:154)
==446768==    by 0x4009061: gi_amlsblk_hash_payload (amlsblk.c:295)
==446768==    by 0x40050B1: gi_bl3_sign_img (bl3.c:52)
==446768==    by 0x40028B6: gi_sign_img (main.c:189)
==446768==    by 0x40032ED: main (main.c:493)
==446768==  Address 0x5105710 is 0 bytes after a block of size 512 alloc'd
==446768==    at 0x485F7A8: malloc (vg_replace_malloc.c:446)
==446768==    by 0x4008D38: gi_amlsblk_hash_payload (amlsblk.c:258)
==446768==    by 0x40050B1: gi_bl3_sign_img (bl3.c:52)
==446768==    by 0x40028B6: gi_sign_img (main.c:189)
==446768==    by 0x40032ED: main (main.c:493)
...
==446768== Syscall param write(buf) points to uninitialised byte(s)
==446768==    at 0x4FFE53D: write (write.c:26)
==446768==    by 0x40081C7: gi_amlsblk_write_blk (amlsblk.c:70)
==446768==    by 0x40087E6: gi_amlsblk_flush_data (amlsblk.c:166)
==446768==    by 0x40050FB: gi_bl3_sign_img (bl3.c:60)
==446768==    by 0x40028B6: gi_sign_img (main.c:189)
==446768==    by 0x40032ED: main (main.c:493)
==446768==  Address 0x1ffefffbe9 is on thread 1's stack
==446768==  in frame #3, created by gi_bl3_sign_img (bl3.c:24)
```