Add a new build option which allows creating reproducible output
If gxlimage is compiled with make REPRODUCIBLE=1 or with
-Dreproducible=true and if SOURCE_DATE_EPOCH is set at runtime, then the
created image will be bit-by-bit reproducible for the same versions of
SOURCE_DATE_EPOCH. This replaces 16 bytes which were filled with values
from /dev/random with a string derived from the SOURCE_DATE_EPOCH value.

Support for meson.options was only added in meson 1.1, so this also adds
a minimum meson version requirement into meson.build.